Added `rows` parameter to `select()`
Added `rows` parameter to `select()` so that the number of rows can be specified. This is useful because the default is 10 and we have cases where we want 0 because we only want the number of records found or we specify the number of rows so that we can get all the records.