Modified `delete()` to list all records to be deleted instead of the first ten
Modified `delete()` to list all records to be deleted instead of the first ten. Solr returns the first ten records be default, so I changed the initial query to return no records because we are only interested in the number of records matching the query. Then when listing the matching records, we make a second query, but this time set rows to the number of records and only return the needed field for performance.