Modified `add()` to add all records at once rather than individually
Modified `add()` to add all records at once rather than individually in order to speed up the tool. This does mean that it can be difficult to track down errors in an update request because all the records are bundled together. We may need to add an option to upload records individually for debug.