Removed `urljoin()` and `urlencode()` from `solr.py`
Removed `urljoin()` and `urlencode()` from `solr.py` because they are no longer needed. Previously I thought I would need to calculate the exact URI size as to maximize a query, but it proved too difficult so I changed the max Solr URI size to 7KB when 8KB is the theoretical max which seems to avoid any issues.

Also removed the unneeded logic from `schema.py` related to this.