Change records from list to dict
Change records from list to dict in manage.py so that we can use `dc_identifier_s` as a UID and reference the record later in processing. For example, this will be useful when warning about duplicate UIDs. We will be able to log the file name of a given UID rather than logging the UID itself.