Allow user to upload if no errors but has warnings, added RecordSet class, moved log writing from `shema.py` to `manage.py`
Allow user to upload if there are no errors but there are warnings. For instance, if there is a warning for duplicate UIDs, a user may choose to ignore the warning and upload anyway.

Added concept of record set to store a set records. Record sets have a method to log errors attached to the records the set contains.

Errors and warnings are now logged in `manage.py` instead of in `schema.py`. `schema.py` will add the error to the record, but it is up to a user to log it.